+++ Source/SourceKittenFrameworkTests/SwiftDocsTests.swift
@@ -10,24 +10,31 @@ import Foundation
 import SourceKittenFramework
 import XCTest
 
+func compareJSONStringWithFixturesName(name: String, jsonString: String) {
+    func jsonValue(jsonString: String) -> AnyObject {
+        let data = jsonString.dataUsingEncoding(NSUTF8StringEncoding)!
+        let result = try! NSJSONSerialization.JSONObjectWithData(data, options: [])
+        return (result as? NSDictionary) ?? (result as! NSArray)
+    }
+    let firstValue = jsonValue(jsonString)
+    let secondValue = jsonValue(File(path: fixturesDirectory + name + ".json")!.contents)
+    let message = "output should match expected fixture"
+    if let firstValue = firstValue as? NSDictionary, secondValue = secondValue as? NSDictionary {
+        XCTAssertEqual(firstValue, secondValue, message)
+    } else if let firstValue = firstValue as? NSArray, secondValue = secondValue as? NSArray {
+        XCTAssertEqual(firstValue, secondValue, message)
+    } else {
+        XCTFail("output didn't match fixture type")
+    }
+}
+
 func compareDocsWithFixturesName(name: String) {
     let swiftFilePath = fixturesDirectory + name + ".swift"
-    let jsonFilePath  = fixturesDirectory + name + ".json"
-
     let docs = SwiftDocs(file: File(path: swiftFilePath)!, arguments: ["-j4", swiftFilePath])
 
     let escapedFixturesDirectory = fixturesDirectory.stringByReplacingOccurrencesOfString("/", withString: "\\/")
-    let comparisonString = docs.description.stringByReplacingOccurrencesOfString(escapedFixturesDirectory, withString: "")
-
-    func docsObject(docsString: String) -> NSDictionary {
-        return try! NSJSONSerialization.JSONObjectWithData(docsString.dataUsingEncoding(NSUTF8StringEncoding)!, options: []) as! NSDictionary
-    }
-
-    XCTAssertEqual(
-        docsObject(comparisonString),
-        docsObject(File(path: jsonFilePath)!.contents),
-        "should generate expected docs for Swift file"
-    )
+    let comparisonString = String(docs).stringByReplacingOccurrencesOfString(escapedFixturesDirectory, withString: "")
+    compareJSONStringWithFixturesName(name, jsonString: comparisonString)
 }
 
 class SwiftDocsTests: XCTestCase {